fix(modes): share unlocked track selection
